use int64 for ids for avoiding possible change afterwards

+++ example/adgroup_create/main.go
@@ -11,8 +11,8 @@ import (
 )
 
 func main() {
-	orgID := 1405310
-	campaignID := 12312323
+	orgID := int64(1405310)
+	campaignID := int64(12312323)
 	client, err := searchads.NewClient(nil, "../cert.pem", "../cert.key", &orgID)
 	if err != nil {
 		log.Fatalf("Client error: %s", err)
