+++ example/campaign_negative_keyword_list/main.go
@@ -10,7 +10,7 @@ import (
 )
 
 func main() {
-	campaignID := 242933727
+	campaignID := int64(242933727)
 	client, err := searchads.NewClient(nil, "../cert.pem", "../cert.key", nil)
 	if err != nil {
 		log.Fatalf("Client error: %s", err)
